split useResponseNavigation
